Update on 01/20/16
Hello

This is the first update for MeteorTodoApp. 98% of the code in the
repository is from the Meteor Todo App tutorial that I am currently
going through. I did change some things like the background color and
the name of the Todo list, but besides that it’s the same as in the
tutorial. Check out the tutorial to make your own Todo app!

Thank You and God Bless!

—Jonathan Bruce